Fix markers that disappeared
Move the data formatting to the container component
Only move the map once.
